% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{add_download}
\alias{add_download}
\title{Register a new download}
\usage{
add_download(x, label, writer = write.csv, description = "", ext = "csv",
  create = TRUE, ...)
}
\arguments{
\item{x}{An object that should be made available for download. If this is
a character vector of length one it is interpreted as the name of the object.}

\item{label}{A short, descriptive label}

\item{writer}{A function used to create the downloadable file from \code{x}.
This function needs to take at least two arguments, the first of which should
be the R object that is to be written to a file and the second of which should
be the file name (without path).}

\item{description}{A longer description.}

\item{ext}{Extension to use for the file.}

\item{create}{Logical indicating wether the downloadable file should be created
immediately.}

\item{...}{Additional arguments are passed to \code{writer}.}
}
\value{
Called for its side effect.
}
\description{
Register a new download
}
\details{
If \code{create} is \code{FALSE} the downloadable file can be created
later by a call to \code{create_download}. This allows objects to be registered
for download when they are created while ensuring that they are not written to
a file until any possible changes to their data have occured.
}
\author{
Peter Humburg
}

