% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{copy_dependencies}
\alias{copy_dependencies}
\title{Copy dependencies to working directory}
\usage{
copy_dependencies(deps)
}
\arguments{
\item{deps}{List of dependencies.}
}
\value{
Called for its side effect.
}
\description{
Output of child documents is copied to the working directory of the main document and installs
a \code{dependencies} package option.
}
\author{
Peter Humburg
}
