% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{create_download}
\alias{create_download}
\title{Create a file for download}
\usage{
create_download(x, download)
}
\arguments{
\item{x}{R object to be saved for download}

\item{download}{Object of class \code{Download}. This may be omitted if
\code{x} was registered for download earlier.}
}
\value{
Called for its side effect.
}
\description{
Create a file for download
}
\author{
Peter Humburg
}

