% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Dependency}
\alias{Dependency}
\title{Create objects of class \code{Dependency}}
\usage{
Dependency(label, document, source, title, cache, ...)
}
\arguments{
\item{label}{The label associated with the document.}

\item{document}{Path to the output document. This should be
relative to the location of the main document.}

\item{source}{Path to source (i.e. RMarkdown) file used to generate \code{document}.}

\item{title}{The title of the document.}

\item{cache}{Path to cache directory. If this is missing an attempt will be made to guess
the location if \code{source} is present.}

\item{...}{Additional arguments are ignored.}
}
\value{
An object of class \code{Dependency}. Such objects have fields
 \item{label}{The label used to refer to this document.}
 \item{document}{The path to the output document.}
 \item{source}{The source document used to generate the output (may be \code{NULL}).}
 \item{title}{The document title. If no title is available this will be 'Untitled'}
 \item{cache}{Location of the cache directory.}
 \item{files}{Path to directory containing additional files associated with the dependency.}
}
\description{
Objects of the class are used to store information
about child docxuments.
}
\details{
The \code{source} and \code{title} arguments may be omitted. If
no \code{title} is provided an attempt is made to extract it from the header
of the source file or, if that is unavailable (or no title was found),
the output document (but currently only for html output).
}

