% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{dependson_opts_hook}
\alias{dependson_opts_hook}
\title{Dependency processing}
\usage{
dependson_opts_hook(options)
}
\arguments{
\item{options}{List of chunk options}
}
\value{
Updated list of options
}
\description{
Hooks for use with knitr to facilitate handling of inter-file dependencies.
}
\author{
Peter Humburg
}

