% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Download}
\alias{Download}
\title{Create objects of class \code{Download}}
\usage{
Download(writer, file_name, label, description, ...)
}
\arguments{
\item{writer}{A function to write the downloadable content to a file. Its first
argument should be the R object to be written and the second argument the file name.}

\item{file_name}{Name of output file, not including any path information.}

\item{label}{A short, descriptive label to identify the download.}

\item{description}{A longer description of the downloadable content.}

\item{...}{Additional arguments are passed to \code{writer}.}
}
\value{
An object of class \code{Download}
}
\description{
Objects of this class hold the information necessary to create output files
containing data and results from the analysis as well as the links necessary
to access them from the HTML output.
}
\note{
This class stores all the information necessary to create downloadable
files, but creating an instance of this class will not create the file.
Use \code{add_download} to add Instances of this class to the global list
of downloads. All downloads registered in this way will be created automatically
after the rest of the document has been processed and a section with download
links is added to the appendix. Download links can be created manually through
calls to \code{create_download}.
}
\author{
Peter Humburg
}
