% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_link}
\alias{download_link}
\title{Create download link}
\usage{
download_link(download, text = download$label, format = c("markdown",
  "html", "table"))
}
\arguments{
\item{download}{An object of class \code{Download}.}

\item{text}{Character vector of length 1 providing the text to display for the link.}

\item{format}{Output format for the link.}
}
\value{
Either a string providing the link in the requested format, or, for
\code{format='table'}, a data frame with a single row and two columns.
}
\description{
Create download link
}

