% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_yaml}
\alias{eval_yaml}
\title{Evaluate R code in YAML fields}
\usage{
eval_yaml(x, data, ...)
}
\arguments{
\item{x}{YAML meta data to process.}

\item{data}{Data to use when evaluating R expressions.}

\item{...}{Further arguments, currently ignored.}
}
\value{
A list of the same shape as \code{x} with all inline R code chunks evaluated.
}
\description{
Evaluate R code in YAML fields
}

