% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_yaml}
\alias{extract_yaml}
\alias{rmd_short_title}
\alias{rmd_title}
\title{Processing YAML metadata}
\usage{
extract_yaml(input, ...)

rmd_title(input, ...)

rmd_short_title(input, ...)
}
\arguments{
\item{input}{Name of input file.}

\item{...}{Additional arguments are ignored.}
}
\value{
\code{extract_yaml} returns a named list with the
contents of the yaml metadata block.

The \code{rmd_*} functions return character vector of
length one containing the value of the corresponding field.
}
\description{
Processing YAML metadata

Extract title from a Markdown document's yaml metadata
}
\details{
\code{rmd_short_title} tries to infer
a suitable short title for a document if it isn't
provided explicitly in the header by using the part
of the title preceding the first punctuation mark
(if any).
}
\author{
Peter Humburg
}

