% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{figRef}
\alias{figRef}
\title{Figure cross-references}
\usage{
figRef(label, caption, prefix = options("figcap.prefix"),
  sep = options("figcap.sep"),
  prefix.highlight = options("figcap.prefix.highlight"))
}
\arguments{
\item{label}{Figure label.}

\item{caption}{Figure caption.}

\item{prefix}{Fixed part of the printed figure label. Defaults to 'Figure'.}

\item{sep}{Separator to use between printed figure label and caption.}

\item{prefix.highlight}{Markdown code the figure label should be wrapped in.
Allows the label to be displayed in bold or italics.}
}
\value{
If the \code{caption} argument is present a string combining the (computed)
figure label with the caption. Otherwise a (markdown formatted) link to the
figure is returned.
}
\description{
Define figure labels and reference them in the text to reference the figure.
Figures are numbered automatically.
}
\details{
Typically this function only needs to be called explicitly to refer
to figures in the text. The call to set the label and generate the appropriately
modified caption is issued automatically when a code chunk with the \code{fig.cap}
option is encountered. In that case the label used in the reference should be the
label of the code chunk that generated the figure. Note that this means code chunks
that generate figures have to be named.

Figure reference can occur at any point in the text. It is not strictly necessary to
define a label before it is referenced. However, figures are numbered in the order
they are first mentioned and this can lead to figures appearing to be out of order.
}
\examples{
options(figcap.prefix='Figure')
options(figcap.prefix.highlight='**')

figRef('foo', 'A test caption')
figRef('foo')
}

