% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{fig.cap_chunk_hook}
\alias{fig.cap_chunk_hook}
\alias{fig.cap_opts_hook}
\title{Special processing for figure chunks}
\usage{
fig.cap_chunk_hook(before, options, envir)

fig.cap_opts_hook(options)
}
\arguments{
\item{before}{Logical indicating whether the hook was called befere or
after the chunk has been processed.}

\item{options}{List of chunk options.}

\item{envir}{Environment in which the chunk is evaluated.}
}
\value{
In the case of \code{fig.cap_chunk_hook} markup used to wrap the figure is returned.

\code{fig.cap_opts_hook} returns a list of chunk options
adjusted for the requested figure format.
}
\description{
These hooks are intended for use with knitr. There is usually no need
to call them directly.
}
\note{
These hooks are intended for chunks with the \code{fig.cap} option and this
is assumed to be present.
}
\author{
Peter Humburg
}

