% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{figureOptions}
\alias{figureOptions}
\title{Manage chunk options for different figure formats}
\usage{
figureOptions(..., format)
}
\arguments{
\item{...}{Named arguments corresponding to knitr chunk options.}

\item{format}{character vector of length one indicting the format
for which options should be set.}
}
\description{
Manage chunk options for different figure formats
}
\author{
Peter Humburg
}
