% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_title}
\alias{html_title}
\title{Extract title from an HTML document}
\usage{
html_title(input, ...)
}
\arguments{
\item{input}{Name of input file.}

\item{...}{Additional arguments are ignored.}
}
\description{
Extract title from an HTML document
}
\author{
Peter Humburg
}
