% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{interactiveFig}
\alias{interactiveFig}
\title{Set figure related chunk options for interactive figures}
\usage{
interactiveFig(fig.width = 8, fig.height = 6, out.width = "800px",
  out.height = "600px", ...)
}
\arguments{
\item{fig.width}{Figure width in inches}

\item{fig.height}{Figure height in inches}

\item{out.width}{Output width for figure in pixels}

\item{out.height}{Output height for figure in pixels}

\item{...}{Additional knitr chunk options}
}
\value{
A list with the previous set of options is returned invisibly.
}
\description{
Set figure related chunk options for interactive figures
}
\author{
Peter Humburg
}

