% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{load_dependencies}
\alias{load_dependencies}
\title{Load results from child documents}
\usage{
load_dependencies(deps, opts)
}
\arguments{
\item{deps}{List of dependencies}

\item{opts}{\code{knitr} options}
}
\value{
Called for its side effect.
}
\description{
Load results from child documents
}
\author{
Peter Humburg
}

