% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_ref_links}
\alias{make_ref_links}
\title{Convert links in a table column to reference links}
\usage{
make_ref_links(data, links = 1, format = c("markdown", "html"))
}
\arguments{
\item{data}{A \code{data.frame} or an object that can be coerced to one.}

\item{links}{Numerical vector of column indices indicating columns that
contain links in need of conversion.}

\item{format}{Format used to specify links in the input.}
}
\value{
A \code{data.frame} with links in the indicated columns replaced
with reference links.
}
\description{
Convert links in a table column to reference links
}
\details{
It is assumed that each entry consists either of a single link
or doesn't contain any link. If a link is identified the resulting reference
link will replace the entire entry. Entries without links will be left unchanged.

Regardles of the input format of the links, the output will always
have markdown reference links.
}
\author{
Peter Humburg
}

