% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_list}
\alias{merge_list}
\title{Merge two named lists}
\usage{
merge_list(x, y)
}
\arguments{
\item{x}{Base list}

\item{y}{Additional entries to be merged into \code{x}}
}
\value{
A list with entries from \code{x} and \code{y}.
}
\description{
Merge two named lists
}
\note{
Entries in \code{x} that are also present in \code{y}
will use the values from \code{y}.
}
