% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-or}
\alias{\%||\%}
\alias{null-or}
\title{Replace null values with default}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Variable to be tested.}

\item{y}{Default to use if \code{x} is \code{NULL}.}
}
\value{
If \code{x} is \code{NULL}, \code{y} is returned, otherwise \code{x}.
}
\description{
Replace null values with default
}
\author{
Peter Humburg
}

