% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{params2deps}
\alias{params2deps}
\title{Extract dependency information from YAML header}
\usage{
params2deps(deps)
}
\arguments{
\item{deps}{List of dependencies}
}
\value{
A list of \code{Dependency} objects
}
\description{
Extract dependency information from YAML header
}
\author{
Peter Humburg
}
