% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{printMD}
\alias{printMD}
\alias{printMD.Dependency}
\alias{printMD.default}
\alias{printMD.double}
\alias{printMD.integer}
\title{Pretty printing of R objects for Markdown output}
\usage{
printMD(x, ...)

\method{printMD}{default}(x, ...)

\method{printMD}{double}(x, big.mark = ",", ...)

\method{printMD}{integer}(x, big.mark = ",", ...)

\method{printMD}{Dependency}(x, format = c("markdown", "html", "reference",
  "md reference"), ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Additional arguments}

\item{big.mark}{Separator used to mark intervals before the decimal point.}

\item{format}{Format type to use.}
}
\value{
String representation of x formatted for printing.
}
\description{
Pretty printing of R objects for Markdown output
}
\examples{
printMD(head(cars))
printMD(10000)
printMD(10000L)
}
\author{
Peter Humburg
}

