% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmd_title}
\alias{rmd_title}
\title{Extract title from a Markdown document's yaml metadata}
\usage{
rmd_title(input, ...)
}
\arguments{
\item{input}{Name of input file.}

\item{...}{Additional arguments are ignored.}
}
\description{
Extract title from a Markdown document's yaml metadata
}
\author{
Peter Humburg
}

