% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{screenFig}
\alias{screenFig}
\title{Set figure related chunk options for static figures}
\usage{
screenFig(fig.width = 8, fig.height = 8, dpi = 300, ...)
}
\arguments{
\item{fig.width}{Figure width in inches}

\item{fig.height}{Figure height in inches}

\item{dpi}{Figure resolution in dots per inch}

\item{...}{Additional knitr chunk options}
}
\value{
A list with the previous set of options is returned invisibly.
}
\description{
Set figure related chunk options for static figures
}
\author{
Peter Humburg
}
