% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup}
\alias{setup}
\title{Setup environment for Multi-document report}
\usage{
setup(params)
}
\arguments{
\item{params}{List of document parameters (as provided in the YAML header)}
}
\value{
Called for its side effect.
}
\description{
Setup environment for Multi-document report
}
\author{
Peter Humburg
}

