% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_size}
\alias{valid_size}
\title{Create valid bootstrap size class
#' @importFrom stringr str_detect}
\usage{
valid_size(size, size_class = "md")
}
\description{
Create valid bootstrap size class
#' @importFrom stringr str_detect
}

