% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{workflow}
\alias{workflow}
\title{Extract workflow description from Rmd files}
\usage{
workflow(main_doc, prefix = file.path(root_dir, "cwl",
  tools::file_path_sans_ext(basename(main_doc))), root_dir = getwd())
}
\arguments{
\item{main_doc}{Path to main document of the report.}

\item{prefix}{Prefix for generated CWL files.}

\item{root_dir}{Use file paths relative to this directory. Set to \code{'/'}
to use absolute paths.}
}
\description{
Parse the documents of a multi-part report to extract the dependencies
between them. This information is then encoded in a
\href{http://www.commonwl.org/v1.0/}{CWL} workflow that
can be used to generate the full report with a compatible workflow
management system (a comprehensive list is available
\href{https://github.com/common-workflow-language/common-workflow-language#implementations}{online}).
Through this mechanism it is possible to take advantage
of parallel processing of documents, where the structure of the report allows.
}
\details{
By default all paths to input files, i.e. the Rmarkdown files
comprising the report, are reported relative to the current working directory.
A different root directory can be chosen by providing an appropriate (absolute)
path as \code{root_dir}. If absolute paths are desired in the output \code{root_dir}
can be set to \code{'/'} but note that this makes the resulting job parameters
less portable.
}
\author{
Peter Humburg
}

