% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{figRef}
\alias{figRef}
\alias{tabRef}
\title{Figure and Table cross-references}
\usage{
figRef(label, caption, target, prefix = options("figcap.prefix"),
  sep = options("figcap.sep"),
  prefix.highlight = options("figcap.prefix.highlight"))

tabRef(label, caption, prefix = options("tabcap.prefix"),
  sep = options("tabcap.sep"),
  prefix.highlight = options("tabcap.prefix.highlight"))
}
\arguments{
\item{label}{Identifying label.}

\item{caption}{Caption to display.}

\item{target}{An object of class \code{Dependency}.}

\item{prefix}{Fixed part of the printed label. Defaults to 'Figure' for figures and to
'Table' for tables.}

\item{sep}{Separator to use between printed label and caption.}

\item{prefix.highlight}{Markdown code the figure label should be wrapped in.
Allows the label to be displayed in bold or italics.}
}
\value{
If the \code{caption} argument is present a string combining the (computed)
figure label with the caption. Otherwise a (markdown formatted) link to the
figure is returned.
}
\description{
Define figure and table labels and reference them in the text to create automatic
reference the corresponding table or figure. Numbers are assigned automatically.
}
\details{
Typically \code{figRef} only needs to be called
explicitly to refer to figures in the text. The call to set the label and generate
the appropriately modified caption is issued automatically when a code chunk with
the \code{fig.cap} option is encountered. In that case the label used in the reference
should be the label of the code chunk that generated the figure. Note that this means
code chunks that generate figures have to be named.

Reference can occur at any point in the text. It is not strictly necessary to
define a label before it is referenced. However, numbering is determined by the order
in which labels are first encountered and this can lead to figures or tables appearing
to be out of order.

It is possible to refer to figures in other documents by supplying a \code{Dependency}
object to argument \code{target}. This will generate a reference of the form
"Short Title, Figure Label", e.g. "Methods, Figure 1". It is an error to provide
a \code{target} as well as a \code{caption}.
}
\examples{
options(figcap.prefix='Figure')
options(figcap.prefix.highlight='**')

figRef('foo', 'A test caption')
figRef('foo')
}
\author{
Peter Humburg
}

