% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{fig.cap_chunk_hook}
\alias{fig.cap_chunk_hook}
\alias{fig.cap_opts_hook}
\alias{fig.height_opts_hook}
\alias{fig.width_opts_hook}
\alias{tab.cap_chunk_hook}
\alias{tab.cap_opts_hook}
\title{Special processing for figure and table chunks}
\usage{
fig.cap_chunk_hook(before, options, envir)

tab.cap_chunk_hook(before, options, envir)

fig.cap_opts_hook(options)

fig.width_opts_hook(options)

fig.height_opts_hook(options)

tab.cap_opts_hook(options)
}
\arguments{
\item{before}{Logical indicating whether the hook was called befere or
after the chunk has been processed.}

\item{options}{List of chunk options.}

\item{envir}{Environment in which the chunk is evaluated.}
}
\value{
The chunk hooks produce markup that adds anchors to enable direct links to the
table or figure and add s the caption where required.

\code{fig.cap_opts_hook} returns a list of chunk options
adjusted for the requested figure format.

\code{fig.width_opts_hook} adjusts figure width for print and screen
figures.

\code{fig.height_opts_hook} adjusts figure height for print and screen
figures.

\code{tab.cap_opts_hook} returns a list of chunk options with
the \code{tab.cap} option augmented for automatic table numbering.
}
\description{
These hooks are intended for use with knitr. There is usually no need
to call them directly.
}
\note{
These hooks are intended for chunks with the \code{fig.cap} (for figures) and
\code{tab.cap} (for tables) option and this is assumed to be present.
}
\author{
Peter Humburg
}

