% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{interactiveFig}
\alias{interactiveFig}
\title{Set figure related chunk options for interactive figures}
\usage{
interactiveFig(out.width = "700px", out.height = "600px", ...)
}
\arguments{
\item{out.width}{Output width for figure in pixels}

\item{out.height}{Output height for figure in pixels}

\item{...}{Additional knitr chunk options}
}
\value{
A list with the previous set of options is returned invisibly.
}
\description{
Set figure related chunk options for interactive figures
}
\author{
Peter Humburg
}

