% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_dependency}
\alias{load_dependencies}
\alias{update_dependency}
\title{Load results from child documents}
\usage{
update_dependency(dep)

load_dependencies(deps)
}
\arguments{
\item{deps}{List of objects of class \code{Dependency}}
}
\value{
Called for its side effect.
}
\description{
Load results from child documents
}
\author{
Peter Humburg
}

