% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_report.R
\name{multi_document}
\alias{multi_document}
\title{Writing Complex Scientific Reports in R}
\usage{
multi_document(theme = NULL, highlight = NULL, pandoc_args = NULL,
  fig_format = c("screen", "print"), fig_download = TRUE,
  tab_download = TRUE, ...)
}
\arguments{
\item{theme}{Visual theme to use for styling of the html output}

\item{highlight}{Visual style to use for Syntax highlighting}

\item{pandoc_args}{Arguments to be passed to pandoc}

\item{fig_format}{Default format(s) for figures}

\item{fig_download}{Logical indicating whether a download link should be added to
figure captions.}

\item{tab_download}{Logical indicating whether a download link should be added to
table captions.}

\item{...}{Additional arguments are passed to rmarkdown::html_document}
}
\description{
An R Markdown document format for multi-part reports.
The *Multi-part Report* template included in the package provides a detailed
example of its use.
}

