% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMD}
\alias{plotMD}
\title{Create plots for inclusion in RMarkdown reports}
\usage{
plotMD(fig, format = c("screen", "interactive", "print"))
}
\arguments{
\item{fig}{A ggplot2 plot for processing.}

\item{format}{A character vector of length one describing the desired
output format.}
}
\value{
Depending on the value of \code{format} either a ggplot2 object
 (possibly the same as \code{fig}) or, if \code{format = 'interactive'}
 a \code{plotly} plot.
}
\description{
Create plots for inclusion in RMarkdown reports
}
\author{
Peter Humburg
}
\seealso{
\code{\link[plotly]{ggplotly}}
}

