% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{printFig}
\alias{printFig}
\title{Set figure related chunk options for print figures}
\usage{
printFig(fig.width = 8, fig.height = 8, ...)
}
\arguments{
\item{fig.width}{Figure width in inches}

\item{fig.height}{Figure height in inches}

\item{...}{Additional knitr chunk options}
}
\value{
A list with the previous set of options is returned invisibly.
}
\description{
Set figure related chunk options for print figures
}
\author{
Peter Humburg
}

